+++ packages/jest-message-util/src/index.js
@@ -64,6 +64,12 @@ const STACK_PATH_REGEXP = /\s*at.*\(?(\:\d*\:\d*|native)\)?/;
 const EXEC_ERROR_MESSAGE = 'Test suite failed to run';
 const ERROR_TEXT = 'Error: ';
 
+const indentAllLines = (lines: string, indent: string) =>
+  lines
+    .split('\n')
+    .map(line => (line ? indent + line : line))
+    .join('\n');
+
 const trim = string => (string || '').trim();
 
 // Some errors contain not only line numbers in stack traces
@@ -84,10 +90,7 @@ const getRenderedCallsite = (
     {highlightCode: true},
   );
 
-  renderedCallsite = renderedCallsite
-    .split('\n')
-    .map(line => MESSAGE_INDENT + line)
-    .join('\n');
+  renderedCallsite = indentAllLines(renderedCallsite, MESSAGE_INDENT);
 
   renderedCallsite = `\n${renderedCallsite}\n`;
   return renderedCallsite;
@@ -127,10 +130,8 @@ export const formatExecError = (
     message = separated.message;
   }
 
-  message = message
-    .split(/\n/)
-    .map(line => MESSAGE_INDENT + line)
-    .join('\n');
+  message = indentAllLines(message, MESSAGE_INDENT);
+
   stack =
     stack && !options.noStackTrace
       ? '\n' + formatStackTrace(stack, config, options, testPath)
@@ -311,10 +312,7 @@ export const formatResultsErrors = (
             formatStackTrace(stack, config, options, testPath),
           ) + '\n';
 
-      message = message
-        .split(/\n/)
-        .map(line => MESSAGE_INDENT + line)
-        .join('\n');
+      message = indentAllLines(message, MESSAGE_INDENT);
 
       const title =
         chalk.bold.red(